pyproject.toml
[project]
name = "procksy"
readme = "README.md"
license = {file = "LICENSE"}
keywords = ["socks", "proxy", "filtering"]
description = "A minimal SOCKS proxy with basic auth and filtering"
requires-python = ">=3.10"
dynamic = ["version"]
authors = [
  { name="CERT-EDF", email="cert@edf.fr" },
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Utilities",
    "Topic :: Security",
]
dependencies = [
    "rich~=13.4",
    "construct~=2.10",
    "argon2-cffi~=23.1",
]


[project.urls]
"Homepage" = "https://github.com/cert-edf/procksy"
"Repository" = "https://github.com/cert-edf/procksy"
"Bug Tracker" = "https://github.com/cert-edf/procksy/issues"


[project.scripts]
procksy = "procksy.main:app"


[build-system]
requires = ["setuptools>=68", "setuptools_scm>=8.0", "wheel"]
build-backend = "setuptools.build_meta"


[tool.setuptools.packages.find]
where = ["."]
include = ["procksy", "procksy.*"]


[tool.setuptools_scm]
version_file = "procksy/__version__.py"


[tool.black]
line-length = 79
skip-string-normalization = true
